tst_map_keepgrab: restore test to conform to expected behavior
Change the test to conform to recent changes in QtQuick related
to Flickable and grabbing. The changes ensures that the user
needs to end an ongoing flicking session before he can start to
interact with a child. From testing native applications and
web pages, this is how it's supposed to work, and also how
Flickable used to work until Qt 6.2. This behavior is therefore
restored.

Task-number: QTBUG-34881
Change-Id: Id30e265a8f2500ad49418b0769d1cb492c08e49b
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>