Fix tile provider resolution getting stuck on disabled providers
When encountering a provider with Enabled=false that has a built-in
valid default we ended up here without submitting any of the result
signals. That resulted in the provider resolution process getting stuck
and no tiles being fetched.

This can be reproduced with using https://autoconfig.kde.org/qtlocation/
as provider repository, instead of Qt's default.

Pick-to: 6.9 6.8
Change-Id: I03a894adeb6c9983f77dad1a5b504b615746d5bd
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>